+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.0.2 (2025-10-19)
+
+### 🚀 Features
+
+- **react-img-mapper:** Added ESM support.
+
+### 🩹 Fixes
+
+- **react-img-mapper:** Required `ref` issue fixed.
+
 ## 2.0.1 (2025-10-18)
 
 ### 🚨 Breaking Change
@@ -15,7 +25,7 @@
 
 ### ❤️ Thank You
 
-@sheepysheepy
+- @sheepysheepy
 
 ## 2.0.0 (2025-01-26)